Fix tests resources cleanup for Openshift (#55638) (#344)
When test resources cleaned up in Openshift, it takes time for the
"istio-ingressgateway" and "istio-egressgateway" services to be deleted.

As a result, when the resources are cleaned after one test and the next
test starts, ittend to fail because the services are in deletion state
and failed to be used by the next test.

Check for the services to be deleted completely before ending the test.

The PR is a cherry-pick of:
- https://github.com/istio/istio/pull/54447
- https://github.com/istio/istio/pull/54377 (partially. only clean part)

Signed-off-by: Maxim Babushkin <mbabushk@redhat.com>
Co-authored-by: Maxim Babushkin <mbabushk@redhat.com>