Sync upstream 1.24.5 (#379)
* Automator: update proxy@release-1.24 in istio/istio@release-1.24 (#55632)

* Release 1.24 fix test services reachability and flackiness (#55642)

* Check for reachable ingress in TestGatewayAPI security test (#55028)

When the Ingress is a domain name, like in public cloud, it might take a
bit for the DNS name to become populated.
Set a check to wait for the domain name to become resolvable.

Signed-off-by: Maxim Babushkin <mbabushk@redhat.com>

* Fix Telemetry Tracint flaky tests failures (#55325)

When running the telemetry tests on public cloud platforms, DNS endpoint
resolve could take some time.
- Increase telemetry tracing tests timeout.
- Use the function with the wait and retry approach to get the endpoint
resolvable.

Signed-off-by: Maxim Babushkin <mbabushk@redhat.com>

---------

Signed-off-by: Maxim Babushkin <mbabushk@redhat.com>

* Automator: update proxy@release-1.24 in istio/istio@release-1.24 (#55662)

* Update BASE_VERSION to 1.24-2025-03-25T19-01-20 (#55671)

* [release-1.24] - Fix TestCNIRaceRepair test for Openshift (#55658)

During the TestCNIRaceRepair test execution, the test deletes the CNI
configuration and restarts the deployment.

In Kind environment, the container at some point gets a "Failure" status
with validation exit error code of 126.
The test checks for that error and continues.

But in Openshift, the container gets stuck in init container status
forever and not sets any exit error code.
But in Openshift we are able to check the container stuck/failure reason
by looking at the events of the container.
In Openshift, the container will show the following message in the events -
"err in GetCNIConfigFromFile".
That could be our indication for the right error status and continue the
test.

Set check for both environments.

Signed-off-by: Maxim Babushkin <mbabushk@redhat.com>

* Automator: update proxy@release-1.24 in istio/istio@release-1.24 (#55682)

* Manual 1.24 backport of #55304 (#55396)

* [release-1.24] disable inbound retry policy for gRPC streaming (#55689)

* disable default inbound retry for gRPC

Signed-off-by: Rama Chavali <rama.rao@salesforce.com>

* release notes

Signed-off-by: Rama Chavali <rama.rao@salesforce.com>

---------

Signed-off-by: Rama Chavali <rama.rao@salesforce.com>
Co-authored-by: Rama Chavali <rama.rao@salesforce.com>

* Fix broken IstioCNI config for enabling repair mode (#55703)

Currently, the CNI `REPAIR_ENABLED` flag is mistakenly enabled
based on `.cni.chained` value instead of `.repair.enabled`.
This PR fixes it.

Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>
Co-authored-by: Sridhar Gaddam <sgaddam@redhat.com>

* Automator: update proxy@release-1.24 in istio/istio@release-1.24 (#55706)

* Update BASE_VERSION to 1.24-2025-04-04T19-02-09 (#55809)

* pilot: fix bug in Alias type RDS cache (#55814)

We accounted for this in DependantConfigs() but not Key(). Can reproduce
by adding/removing a Alias service from a Sidecar scope and you will see
its not updated before this PR. With RDS cache off, or
`cachez?clear=true`, or this PR, it is updated.

Co-authored-by: John Howard <john.howard@solo.io>

* Fix tests resources cleanup for Openshift (#55638)

When test resources cleaned up in Openshift, it takes time for the
"istio-ingressgateway" and "istio-egressgateway" services to be deleted.

As a result, when the resources are cleaned after one test and the next
test starts, ittend to fail because the services are in deletion state
and failed to be used by the next test.

Check for the services to be deleted completely before ending the test.

The PR is a cherry-pick of:
- https://github.com/istio/istio/pull/54447
- https://github.com/istio/istio/pull/54377 (partially. only clean part)

Signed-off-by: Maxim Babushkin <mbabushk@redhat.com>

* Do not split root CA, fixing 1.25.1 regression (#55850)

Fixes https://github.com/istio/istio/issues/55793

Also adds istioctl support for the ztunnel side fixes in https://github.com/istio/ztunnel/pull/1510

Co-authored-by: John Howard <john.howard@solo.io>

* Automator: update common-files@release-1.24 in istio/istio@release-1.24 (#55861)

* Automator: update istio/client-go@release-1.24 dependency in istio/istio@release-1.24 (#55863)

* Automator: update proxy@release-1.24 in istio/istio@release-1.24 (#55865)

* Automator: update ztunnel@release-1.24 in istio/istio@release-1.24 (#55883)

* Automator: update proxy@release-1.24 in istio/istio@release-1.24 (#55893)

---------

Signed-off-by: Maxim Babushkin <mbabushk@redhat.com>
Signed-off-by: Rama Chavali <rama.rao@salesforce.com>
Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>
Co-authored-by: Istio Automation <istio-testing-bot@google.com>
Co-authored-by: Maxim Babushkin <mbabushk@redhat.com>
Co-authored-by: Ben Leggett <854255+bleggett@users.noreply.github.com>
Co-authored-by: Rama Chavali <rama.rao@salesforce.com>
Co-authored-by: Sridhar Gaddam <sgaddam@redhat.com>
Co-authored-by: John Howard <john.howard@solo.io>