Add domain resolve to Ambient TestServiceEntrySelectsWorkloadEntry test (#56861)
When executing the TestServiceEntrySelectsWorkloadEntry Ambient test in
a public cloud based cluster, the created ingress is a domain based name
and the test fails to operate with the domain name instead of the ip
address.

Add a domain name resolution to ip based address.

The PR similar to the fix being done for another test:
https://github.com/istio/istio/pull/56784

Signed-off-by: Maxim Babushkin <mbabushk@redhat.com>