Tests fixes into release 1.26 branch (#439)
* Add domain resolve to Ambient TestServiceEntrySelectsWorkloadEntry test (#56861)

When executing the TestServiceEntrySelectsWorkloadEntry Ambient test in
a public cloud based cluster, the created ingress is a domain based name
and the test fails to operate with the domain name instead of the ip
address.

Add a domain name resolution to ip based address.

The PR similar to the fix being done for another test:
https://github.com/istio/istio/pull/56784

Signed-off-by: Maxim Babushkin <mbabushk@redhat.com>

* Add domain resolve to Ambient TestServiceEntryInlinedWorkloadEntry test (#56784)

When executing TestServiceEntryInlinedWorkloadEntry Ambient test in a
public cloud based cluster, the created ingress is a domain based name
and the test fails with the following error:

configuration is invalid: endpoint address
"a6eec2a91721a42b0aa8e140fc6611fc-1873540778.us-east-1.elb.amazonaws.com"
is not a valid IP address

Add a domain name resolution to ip based address.

Signed-off-by: Maxim Babushkin <mbabushk@redhat.com>

* Fix TestMTlsGatewaysWithQUIC test crash on Openshift (#54926)

The TestMTlsGatewaysWithQUIC and TestTlsGatewaysWithQUIC security tests
create a mixed TCP/UDP "istio-ingressgateway" service as part of the
test tear up to test the QUIC feature.

Some cloud platform may throw the following error during creation of
the service with mixed TCP/UDP protocols:

"Error syncing load balancer: failed to ensure load balancer:
mixed protocol is not supported for LoadBalancer".

Check the ingress service events to validate creation of the service.
If the service could not be created due to lack of support from cloud
provider, skip the test.

Signed-off-by: Maxim Babushkin <mbabushk@redhat.com>

---------

Signed-off-by: Maxim Babushkin <mbabushk@redhat.com>