[release-1.27] Fix the proxy image to work with new proxy binary (#453)
The envoy binary is built in a Fedora 41 container, and when it runs on
Rockylinux we get errors like
```
envoy: /lib64/libm.so.6: version `GLIBC_2.38' not found (required by envoy)
```

So we need the sidecar image to also be based on Fedora.

Note that we already diverge a bit from upstream in this area.