Do not create the PDB by default when we don't have enough replicas (#57202)
* Do not create the PDB by default when we don't have enough replicas

Currently used default values are creating the PDB with minAvailable=1 and HPA with minReplicas=1.
The result of that is https://github.com/istio/istio/issues/12602. This PR
makes sure that PDB is not created by default and is created only if
users increase either minReplicas in HPA or replicaCount in the
Deployment.

Fixes https://github.com/istio/istio/issues/12602

Signed-off-by: Filip Brychta <fbrychta@redhat.com>

* Adding/fixing helm unit tests

Signed-off-by: Filip Brychta <fbrychta@redhat.com>

* Fixing manifest generate unit tests

Signed-off-by: Filip Brychta <fbrychta@redhat.com>

* Adding release notes

Signed-off-by: Filip Brychta <fbrychta@redhat.com>

---------

Signed-off-by: Filip Brychta <fbrychta@redhat.com>