Modify version slection in Sail control plane install flow for testing
When we are deploying Sail Operator as our control plane for integration
testing and we're not providing ISTIO_VERSION env var, it takes the
deployment version automatically.
Currently, it takes the last released (upstream) version of Sail
Operator.
But it leads to a situation, where Istio testing executed on "master"
(latest), does not have required suported featured because the version
is not aligned.

Make sure to align the version of Sail Operator being deployed for the
testing.
The flow will use the CONVERTER_BRANCH to install the relevant Sail
OPerator version.
If not defined, it will deploy the latest available to align testing
versions with Istio integration testing flow.

Signed-off-by: Maxim Babushkin <mbabushk@redhat.com>